+++ lib/are_search/searcher/search_body_policy.rb
@@ -8,14 +8,12 @@ module AreSearch
 
         class << self
 
-            def valid?(_es_params)
-                raise NotImplementedError,
-                    "#{name}.valid? を実装してください"
+            def valid?(es_params)
+                raise NotImplementedError, "#{name}.valid? を実装してください"
             end
 
-            def invalid_key?(_key_name)
-                raise NotImplementedError,
-                    "#{name}.invalid_key? を実装してください"
+            def invalid_key?(key_name)
+                raise NotImplementedError, "#{name}.invalid_key? を実装してください"
             end
         end
     end